Progress on #181 and #184.
Still working on the incorrect password dialog for #181.  Maybe another hour or two for that.

Didn't add an "Add Friend" confirmation in #184.  Do we really need it? Was going to ask at scrum meeting.  Lemme know if you want it, it'll be 20 mins work to copy the "remove" confirmation modal.